Formatter and Reducer moved to their dedicated files
